Moved bannedStrings to seperate file, added admins table to schema.
